feat: editing notes
